// src/app/global/configs/routes/api/billing-address.ts
export const enum BILLING_ADDRESS_API {
  getById = 'billingAddress/fetchBy/id/',
  fetchByUserId = 'billingAddress/fetchBy/userId/',
  fetchAll = 'billingAddress/fetch/all',
  create = 'billingAddress/new',
  createList = 'billingAddress/new/list',
  update = 'billingAddress/update',
  deleteById = 'billingAddress/deleteBy/id/',
  deleteList = 'billingAddress/delete/list',
}